move input handling to component
