+++ packages/app/src/cli/services/function/ui/components/Replay/hooks/useFunctionWatcher.ts
@@ -7,8 +7,6 @@ import {FunctionRunFromRunner, ReplayLog} from '../types.js'
 import {exec} from '@shopify/cli-kit/node/system'
 import {AbortController} from '@shopify/cli-kit/node/abort'
 import {useEffect, useState} from 'react'
-import {useInput, useStdin} from '@shopify/cli-kit/node/ink'
-import {handleCtrlC} from '@shopify/cli-kit/node/ui'
 import {useAbortSignal} from '@shopify/cli-kit/node/ui/hooks'
 import {isUnitTest} from '@shopify/cli-kit/node/context/local'
 import {treeKill} from '@shopify/cli-kit/node/tree-kill'
@@ -43,8 +41,6 @@ export function useFunctionWatcher({selectedRun, abortController, app, extension
   const [error, setError] = useState<string | undefined>(undefined)
 
   const {input, export: runExport} = selectedRun.payload
-
-  const {isRawModeSupported: canUseShortcuts} = useStdin()
 
   const [statusMessage, setStatusMessage] = useState(`Watching for changes to ${selectedRun.source}...`)
 
@@ -108,43 +104,16 @@ export function useFunctionWatcher({selectedRun, abortController, app, extension
     }
   }, [input, runExport, app, extension])
 
-  useAbortSignal(abortController.signal, async (err) => {
-    if (err) {
-      setStatusMessage('Shutting down replay watcher because of an error ...')
-    } else {
-      setStatusMessage('Shutting down replay watcher ...')
-      setTimeout(() => {
-        if (isUnitTest()) return
-        treeKill(process.pid, 'SIGINT', false, () => {
-          process.exit(0)
-        })
-      }, 2000)
-    }
+  useAbortSignal(abortController.signal, async () => {
+    setTimeout(() => {
+      if (isUnitTest()) return
+      treeKill(process.pid, 'SIGINT', false, () => {
+        process.exit(0)
+      })
+    }, 2000)
   })
 
-  useInput(
-    (input, key) => {
-      handleCtrlC(input, key, () => abortController.abort())
-
-      const onInput = async () => {
-        try {
-          setError('')
-
-          if (input === 'q') {
-            abortController.abort()
-          }
-          // eslint-disable-next-line no-catch-all/no-catch-all
-        } catch (_) {
-          setError('Failed to handle your input.')
-        }
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      onInput()
-    },
-    {isActive: Boolean(canUseShortcuts)},
-  )
-  return {logs, canUseShortcuts, statusMessage, recentFunctionRuns, error}
+  return {logs, statusMessage, recentFunctionRuns, error}
 }
 
 async function runFunctionRunnerWithLogInput(
